frontend/requirements.txt
streamlit>=1.30.0
requests>=2.30.0


python-dotenv>=1.0.0
